move routes map service to content API coontainer
