+++ packages/core/admin/server/services/index.js
@@ -14,5 +14,4 @@ module.exports = {
   action: require('./action'),
   'api-token': require('./api-token'),
   'project-settings': require('./project-settings'),
-  'content-api': require('./content-api'),
 };
